+++ src/index.tsx
@@ -2,10 +2,24 @@ import { TelemetryDeckContext } from "./telemetrydeck-context";
 import { TelemetryDeckProvider } from "./telemetrydeck-provider";
 import { createTelemetryDeck } from "./create-telemetrydeck";
 import { useTelemetryDeck } from "./use-telemetrydeck";
+import {
+  webBrowserPlugin,
+  webEnvironmentPlugin,
+  webAccessibilityPlugin,
+  webNavigationPlugin,
+  webPwaPlugin,
+  nativePlugin,
+} from "./plugins";
 
 export {
   TelemetryDeckContext,
   TelemetryDeckProvider,
   createTelemetryDeck,
   useTelemetryDeck,
+  webBrowserPlugin,
+  webEnvironmentPlugin,
+  webAccessibilityPlugin,
+  webNavigationPlugin,
+  webPwaPlugin,
+  nativePlugin,
 };